% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestryData-data.R
\docType{data}
\name{ancestryData}
\alias{ancestryData}
\title{Example allele frequency data}
\format{
Chromosome, SNP, base pair, reference and alternate alleles, 
 reference allele frequencies, observed allele frequencies
}
\usage{
data(ancestryData)
}
\description{
reference data is 1000 Genomes and NAM. 
Observed data is from gnomAD:
african/african american, 
american/admixed latinx, 
and other
}
\examples{
data("ancestryData")
}
\keyword{datasets}
